+++ python/xpctl/backends/backend.py
@@ -1,5 +1,21 @@
 from copy import deepcopy
-from baseline.utils import export
+import shutil
+from collections import namedtuple
+import json
+import os
+import numpy as np
+from flask import abort
+
+from baseline.utils import export, write_config_file, unzip_model
+
+from xpctl.swagger_server.models import Experiment as ServerExperiment
+from xpctl.swagger_server.models import Result as ServerResult
+from xpctl.swagger_server.models import ExperimentAggregate as ServerExperimentAggregate
+from xpctl.swagger_server.models import Response as ServerResponse
+from xpctl.swagger_server.models import TaskSummary as ServerTaskSummary
+from xpctl.swagger_server.models import AggregateResult as ServerAggregateResult
+from xpctl.swagger_server.models import AggregateResultValues
+
 
 __all__ = []
 exporter = export(__all__)
@@ -8,6 +24,9 @@ TRAIN_EVENT = 'train_events'
 VALID_EVENT = 'valid_events'
 TEST_EVENT = 'test_events'
 EVENT_TYPES = [TRAIN_EVENT, VALID_EVENT, TEST_EVENT]
+
+
+METRICS_SORT_ASCENDING = ['avg_loss', 'perplexity']
 
 
 @exporter
@@ -19,7 +38,7 @@ class Result(object):
         self.tick_type = tick_type
         self.tick = tick
         self.phase = phase
-
+    
     def get_prop(self, field):
         if field not in self.__dict__:
             raise ValueError('{} does not have a property {}'.format(self.__class__, field))
@@ -348,3 +367,267 @@ class Success(Response):
         self.message = message
         self.response_type = response_type
         self.code = code
+
+
+@exporter
+def log2json(log_file):
+    s = []
+    with open(log_file) as f:
+        for line in f:
+            x = line.replace("'", '"')
+            s.append(json.loads(x))
+    return s
+
+
+@exporter
+def json2log(events, log_file):
+    with open(log_file, 'w') as wf:
+        for event in events:
+            wf.write(json.dumps(event)+'\n')
+
+
+@exporter
+def get_checkpoint(checkpoint_base, checkpoint_store, config_sha1, hostname):
+    if checkpoint_base:
+        model_loc = store_model(checkpoint_base, config_sha1, checkpoint_store)
+        if model_loc is not None:
+            return "{}:{}".format(hostname, os.path.abspath(model_loc))
+        else:
+            raise RuntimeError("model could not be stored, see previous errors")
+
+
+@exporter
+def get_experiment_label(config_obj, task, **kwargs):
+    if kwargs.get('label', None) is not None:
+        return kwargs['label']
+    if 'description' in config_obj:
+        return config_obj['description']
+    else:
+        model_type = config_obj.get('model_type', 'default')
+        backend = config_obj.get('backend', 'tensorflow')
+        return "{}-{}-{}".format(task, backend, model_type)
+
+
+@exporter
+def safe_get(_object, key, alt):
+    val = _object.get(key)
+    if val is None or str(val) is None:
+        return alt
+    return val
+
+
+@exporter
+def store_model(checkpoint_base, config_sha1, checkpoint_store, print_fn=print):
+    checkpoint_base = unzip_model(checkpoint_base)
+    mdir, mbase = os.path.split(checkpoint_base)
+    mdir = mdir if mdir else "."
+    if not os.path.exists(mdir):
+        print_fn("no directory found for the model location: [{}], aborting command".format(mdir))
+        return None
+    
+    mfiles = ["{}/{}".format(mdir, x) for x in os.listdir(mdir) if x.startswith(mbase + "-") or
+              x.startswith(mbase + ".")]
+    if not mfiles:
+        print_fn("no model files found with base [{}] at location [{}], aborting command".format(mbase, mdir))
+        return None
+    model_loc_base = "{}/{}".format(checkpoint_store, config_sha1)
+    if not os.path.exists(model_loc_base):
+        os.makedirs(model_loc_base)
+    dirs = [int(x[:-4]) for x in os.listdir(model_loc_base) if x.endswith(".zip") and x[:-4].isdigit()]
+    # we expect dirs in numbers.
+    new_dir = "1" if not dirs else str(max(dirs) + 1)
+    model_loc = "{}/{}".format(model_loc_base, new_dir)
+    os.makedirs(model_loc)
+    for mfile in mfiles:
+        shutil.copy(mfile, model_loc)
+        print_fn("writing model file: [{}] to store: [{}]".format(mfile, model_loc))
+    print_fn("zipping model files")
+    shutil.make_archive(base_name=model_loc,
+                        format='zip',
+                        root_dir=model_loc_base,
+                        base_dir=new_dir)
+    shutil.rmtree(model_loc)
+    print_fn("model files zipped and written")
+    return model_loc + ".zip"
+
+
+@exporter
+def serialize_experiment_details(exp):
+    if type(exp) == Error:
+        return abort(500, exp.message)
+    train_events = [ServerResult(**r.__dict__) for r in exp.train_events]
+    valid_events = [ServerResult(**r.__dict__) for r in exp.valid_events]
+    test_events = [ServerResult(**r.__dict__) for r in exp.test_events]
+    d = exp.__dict__
+    d.update({'train_events': train_events})
+    d.update({'valid_events': valid_events})
+    d.update({'test_events': test_events})
+    return ServerExperiment(**d)
+
+
+@exporter
+def serialize_get_results(agg_exps):
+    if type(agg_exps) == Error:
+        return abort(500, agg_exps.message)
+    results = []
+    for agg_exp in agg_exps:
+        train_events = [ServerAggregateResult(metric=r.metric,
+                                              values=[AggregateResultValues(k, v) for k, v in r.values.items()])
+                        for r in agg_exp.train_events]
+        valid_events = [ServerAggregateResult(metric=r.metric,
+                                              values=[AggregateResultValues(k, v) for k, v in r.values.items()])
+                        for r in agg_exp.valid_events]
+        test_events = [ServerAggregateResult(metric=r.metric,
+                                             values=[AggregateResultValues(k, v) for k, v in r.values.items()])
+                       for r in agg_exp.test_events]
+        
+        d = agg_exp.__dict__
+        d.update({'train_events': train_events})
+        d.update({'valid_events': valid_events})
+        d.update({'test_events': test_events})
+        results.append(ServerExperimentAggregate(**d))
+    return results
+
+
+@exporter
+def serialize_list_results(exps):
+    if type(exps) == Error:
+        return abort(500, exps.message)
+    results = []
+    for exp in exps:
+        if type(exp) == Error:
+            return Response(**exp.__dict__)
+        train_events = [ServerResult(**r.__dict__) for r in exp.train_events]
+        valid_events = [ServerResult(**r.__dict__) for r in exp.valid_events]
+        test_events = [ServerResult(**r.__dict__) for r in exp.test_events]
+        d = exp.__dict__
+        d.update({'train_events': train_events})
+        d.update({'valid_events': valid_events})
+        d.update({'test_events': test_events})
+        results.append(ServerExperiment(**d))
+    return results
+
+
+@exporter
+def serialize_task_summary(task_summary):
+    if type(task_summary) == Error:
+        return abort(500, task_summary.message)
+    return ServerTaskSummary(**task_summary.__dict__)
+
+
+@exporter
+def serialize_summary(task_summaries):
+    _task_summaries = []
+    for task_summary in task_summaries:
+        if type(task_summary) != Error:  # should we abort if we cant get summary for a task in the database?
+            _task_summaries.append(ServerTaskSummary(**task_summary.__dict__))
+    return _task_summaries
+
+
+@exporter
+def serialize_config2json(config):
+    if type(config) == Error:
+        return abort(500, config.message)
+    return config
+
+
+@exporter
+def serialize_get_model_location(location):
+    return ServerResponse(**location.__dict__)
+
+
+@exporter
+def serialize_post_requests(result):
+    return ServerResponse(**result.__dict__)
+
+
+@exporter
+def deserialize_result(result):
+    return Result(
+        metric=result.metric,
+        value=result.value,
+        tick_type=result.tick_type,
+        tick=result.tick,
+        phase=result.phase
+    )
+
+
+@exporter
+def deserialize_experiment(exp):
+    train_events = [deserialize_result(r) for r in exp.train_events]
+    valid_events = [deserialize_result(r) for r in exp.valid_events]
+    test_events = [deserialize_result(r) for r in exp.test_events]
+    return Experiment(
+        task=exp.task,
+        eid=exp.eid,
+        username=exp.username,
+        hostname=exp.hostname,
+        config=exp.config,
+        exp_date=exp.exp_date,
+        label=exp.label,
+        dataset=exp.dataset,
+        sha1=exp.sha1,
+        version=exp.version,
+        train_events=train_events,
+        valid_events=valid_events,
+        test_events=test_events
+    )
+
+
+@exporter
+def pack_results_in_events(results):
+    d = {}
+    for result in results:
+        if result.tick not in d:
+            d[result.tick] = {result.metric: result.value,
+                              'tick_type': result.tick_type,
+                              'phase': result.phase,
+                              'tick': result.tick
+                              }
+        else:
+            d[result.tick].update({result.metric: result.value})
+    return list(d.values())
+
+
+@exporter
+def experiment_to_put_result_consumable(exp):
+    d = exp.__dict__
+    train_events = pack_results_in_events(exp.train_events)
+    d.pop('train_events')
+    valid_events = pack_results_in_events(exp.valid_events)
+    d.pop('valid_events')
+    test_events = pack_results_in_events(exp.test_events)
+    d.pop('test_events')
+    config = exp.config
+    d.pop('config')
+    task = exp.task
+    d.pop('task')
+    put_result_consumable = namedtuple('put_result_consumable', ['task', 'config_obj', 'events_obj', 'extra_args'])
+    return put_result_consumable(task=task, config_obj=json.loads(config),
+                                 events_obj=train_events+valid_events+test_events,
+                                 extra_args=d)
+
+
+@exporter
+def aggregate_results(resultset, reduction_dim, event_type, num_exps_per_reduction):
+    # TODO: implement a trim method for ExperimentGroup
+    grouped_result = resultset.groupby(reduction_dim)
+    aggregate_fns = {'min': np.min, 'max': np.max, 'avg': np.mean, 'std': np.std}
+    return grouped_result.reduce(aggregate_fns=aggregate_fns, event_type=event_type)
+
+
+@exporter
+def write_experiment(experiment, basedir):
+    eid = str(experiment.eid)
+    basedir = os.path.join(basedir, eid)
+    os.makedirs(basedir)
+    train_events = pack_results_in_events(experiment.train_events)
+    valid_events = pack_results_in_events(experiment.valid_events)
+    test_events = pack_results_in_events(experiment.test_events)
+    json2log(train_events + valid_events + test_events, os.path.join(basedir, '{}-reporting.log'.format(eid)))
+    config = json.loads(experiment.config) if type(experiment.config) is str else experiment.config
+    write_config_file(config, os.path.join(basedir, '{}-config.yml'.format(eid)))
+    d = experiment.__dict__
+    [d.pop(event_type) for event_type in EVENT_TYPES]
+    d.pop('config')
+    write_config_file(d, os.path.join(basedir, '{}-meta.yml'.format(eid)))
