+++ python/xpctl/swagger_server/controllers/xpctl_controller.py
@@ -2,7 +2,7 @@ import connexion
 
 from swagger_server.models.experiment import Experiment  # noqa: E501
 import flask
-from xpctl.backends.dto import serialize_config2json, serialize_experiment_details, serialize_get_model_location, \
+from xpctl.backends.backend import serialize_config2json, serialize_experiment_details, serialize_get_model_location, \
     serialize_get_results, serialize_list_results, serialize_post_requests, deserialize_experiment, serialize_summary, \
     serialize_task_summary
